+++ library/src/assignment_policy.cpp
@@ -258,8 +258,14 @@ bool AssignmentPolicy::ValidOutBuffer(ExecPlan&         execPlan,
         if(BufferIsUnitStride(execPlan, buffer))
         {
             // just check if there's enough space
-            return std::accumulate(nodeLen.begin(), nodeLen.end(), 1, std::multiplies<size_t>())
-                   <= std::accumulate(bufLen.begin(), bufLen.end(), 1, std::multiplies<size_t>());
+            return std::accumulate(nodeLen.begin(),
+                                   nodeLen.end(),
+                                   static_cast<size_t>(1),
+                                   std::multiplies<size_t>())
+                   <= std::accumulate(bufLen.begin(),
+                                      bufLen.end(),
+                                      static_cast<size_t>(1),
+                                      std::multiplies<size_t>());
         }
 
         // NB:
